Drop Datasource/QueryTrait
